# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: InvestFlow Investment Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /positions:
    get:
      operationId: listPositions
      tags: [positions]
      summary: List my positions (cotas)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Positions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"

    post:
      operationId: buyPosition
      tags: [positions]
      summary: Buy strategy position (cota)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PositionInput"
      responses:
        "201":
          description: Position created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        "400":
          description: Insufficient balance or invalid amount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /positions/{id}:
    get:
      operationId: getPosition
      tags: [positions]
      summary: Get position details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Position details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
  /positions/summary:
    get:
      operationId: getPositionsSummary
      tags: [positions]
      summary: Get positions summary by strategy
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Positions summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionsSummary"

  # DEPOSITS
  /deposits:
    get:
      operationId: listDeposits
      tags: [deposits]
      summary: List my deposits
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, confirmed, failed, cancelled]
        - in: query
          name: method
          schema:
            type: string
      responses:
        "200":
          description: Deposits list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"

    post:
      operationId: createDeposit
      tags: [deposits]
      summary: Create deposit request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /deposits/{id}:
    get:
      operationId: getDeposit
      tags: [deposits]
      summary: Get deposit details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /deposits/methods:
    get:
      operationId: getDepositMethods
      tags: [deposits]
      summary: Get available deposit methods and wallet addresses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deposit methods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepositMethod"

  # WITHDRAWALS
...
  /referrals:
    get:
      operationId: getReferralInfo
      tags: [referrals]
      summary: Get referral program info and my referrals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Referral info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralInfo"
  /referrals/tree:
    get:
      operationId: getReferralTree
      tags: [referrals]
      summary: Get referral tree visualization data
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Referral tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralTree"
  /referrals/commissions:
    get:
      operationId: listCommissions
      tags: [referrals]
      summary: List commissions earned
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Commissions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commission"

  # NOTIFICATIONS
...
  /admin/deposits:
    get:
      operationId: adminListDeposits
      tags: [admin]
      summary: Admin list all deposits
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Deposits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositPage"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Position:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        strategyId:
          type: integer
        strategyName:
          type: string
        shares:
          type: number
        purchasePrice:
          type: number
        currentValue:
          type: number
        investedAmount:
          type: number
        yieldAmount:
          type: number
        yieldPercentage:
          type: number
        status:
          type: string
          enum: [active, closed]
        purchasedAt:
          type: string
          format: date-time
      required: [id, userId, strategyId, strategyName, shares, purchasePrice, currentValue, investedAmount, yieldAmount, yieldPercentage, status, purchasedAt]
    PositionInput:
      type: object
      required: [strategyId, amount]
      properties:
        strategyId:
          type: integer
        amount:
          type: number
          minimum: 0.01
    PositionsSummary:
      type: object
      properties:
        totalInvested:
          type: number
        totalCurrentValue:
          type: number
        totalYield:
          type: number
        totalYieldPercentage:
          type: number
        byStrategy:
          type: array
          items:
            $ref: "#/components/schemas/Position"
      required: [totalInvested, totalCurrentValue, totalYield, totalYieldPercentage, byStrategy]

    # DEPOSIT SCHEMAS
    Deposit:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        method:
          type: string
          enum: [pix, usdt_bep20, bitcoin, usdc, bnb]
        amount:
          type: number
        status:
          type: string
          enum: [pending, confirmed, failed, cancelled]
        walletAddress:
          type: ["string", "null"]
        transactionHash:
          type: ["string", "null"]
        qrCodeUrl:
          type: ["string", "null"]
        confirmedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, method, amount, status, createdAt]
    DepositInput:
      type: object
      required: [method, amount]
      properties:
        method:
          type: string
          enum: [pix, usdt_bep20, bitcoin, usdc, bnb]
        amount:
          type: number
          minimum: 1
    DepositMethod:
      type: object
      properties:
        method:
          type: string
        label:
          type: string
        walletAddress:
          type: ["string", "null"]
        qrCodeUrl:
          type: ["string", "null"]
        instructions:
          type: string
        minAmount:
          type: number
        isActive:
          type: boolean
      required: [method, label, instructions, minAmount, isActive]
...
    DepositPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Deposit"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, totalPages]

    # WITHDRAWAL SCHEMAS
...
    ReferralInfo:
      type: object
      properties:
        referralCode:
          type: string
        referralLink:
          type: string
        totalReferrals:
          type: integer
        activeReferrals:
          type: integer
        totalCommissionEarned:
          type: number
        pendingCommission:
          type: number
        commissionRate:
          type: number
        referrals:
          type: array
          items:
            $ref: "#/components/schemas/ReferralMember"
      required: [referralCode, referralLink, totalReferrals, activeReferrals, totalCommissionEarned, pendingCommission, commissionRate, referrals]
    ReferralMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        status:
          type: string
        totalInvested:
          type: number
        commissionGenerated:
          type: number
        joinedAt:
          type: string
          format: date-time
      required: [id, name, email, status, totalInvested, commissionGenerated, joinedAt]
    ReferralTree:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        level:
          type: integer
        children:
          type: array
          items:
            $ref: "#/components/schemas/ReferralTree"
      required: [id, name, email, level, children]
    Commission:
      type: object
      properties:
        id:
          type: integer
        fromUserId:
          type: integer
        fromUserName:
          type: string
        amount:
          type: number
        rate:
          type: number
        level:
          type: integer
        status:
          type: string
          enum: [pending, paid]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, fromUserId, fromUserName, amount, rate, level, status, createdAt]

    # NOTIFICATION SCHEMAS
...