# lib/api-zod/src/generated/types/position.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * InvestFlow Investment Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PositionStatus } from './positionStatus';

export interface Position {
  id: number;
  userId: number;
  strategyId: number;
  strategyName: string;
  shares: number;
  purchasePrice: number;
  currentValue: number;
  investedAmount: number;
  yieldAmount: number;
  yieldPercentage: number;
  status: PositionStatus;
  purchasedAt: Date;
}